svr/requirements.txt
aiofiles==22.1.0
entrance==1.1.17
httptools==0.5.0
multidict==6.0.3
PyYAML==6.0
sanic==22.9.1
sanic-routing==22.8.0
ujson==5.6.0
uvloop==0.17.0
websockets==10.4
